requirements.txt
httpx
requests
fastapi
jinja2
python-dotenv
